+++ Sourcery/CodeGenerated/Coding.generated.swift
@@ -3,7 +3,6 @@
 
 import Foundation
 
-
 extension NSCoder {
 
     @nonobjc func decode(forKey: String) -> String? {